chore(suite): reorganize WalletLayout folder
