+++ packages/suite/src/components/wallet/CoinjoinAccountDiscovery/CoinjoinAccountDiscoveryProgress.tsx
