+++ packages/suite/src/components/wallet/CoinjoinAccountDiscovery/RememberWallet.tsx
