+++ packages/suite/src/components/wallet/CoinjoinAccountDiscovery/lottie/cubes_line.json
