+++ packages/suite/src/components/wallet/CoinjoinAccountDiscovery/lottie/square_stack.json
