+++ packages/suite/src/components/wallet/WalletLayout/AccountBanners/XRPReserve.tsx
