+++ packages/suite/src/components/wallet/WalletLayout/AccountException/AccountException.tsx
