+++ packages/suite/src/components/wallet/WalletLayout/AccountException/DiscoveryEmpty.tsx
