+++ packages/suite/src/components/wallet/WalletLayout/AccountException/DiscoveryFailed.tsx
