+++ packages/suite/src/components/wallet/WalletLayout/AccountTopPanel/AccountNavigation.tsx
