+++ packages/suite/src/components/wallet/WalletLayout/AccountsMenu/AccountItemSkeleton.tsx
