+++ packages/suite/src/components/wallet/WalletLayout/AccountsMenu/AnimationWrapper.tsx
