+++ packages/suite/src/components/wallet/WalletLayout/WalletLayoutHeader.tsx
